+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # VRFSearchTool.py #
 ---
 
+## VRFSearchTool 0.08 (2013-07-29) ##
+* Updated CHANGELOG.md, README.md, TODO.md
+* Minor corrections to code comments
+
+## VRFSearchTool 0.07 (2013-07-27) ##
+* Added error checking for user input.  Also fixed application crashing when
+  accepting input < 2 characters from user.
+* Removed unnecessary debugging commands (no longer needed to trace flow
+  of function calls).
+* Extended spacing for output table to fit longer VRF Names (up to 20 characters)
+
 ## VRFSearchTool 0.06 (2013-07-25) ##
-* Alpha release of the tool.  Still requires some error checking, code cleanup
+* Alpha release of the tool.  Still requires some error checking, code clean-up
   and commenting but otherwise is a working application.
 
 ## VRFSearchTool 0.01 (2013-07-12) ##